Changes in response to PR comments
Configure Inverse Search Action back to showing when Sumatra is available.
Redone InputVerifier in LatexSettingsEditor to just update the ComboBox and SumatraAvailabilityChecker, no focus changes.
Added a file for SumatraAvailabilityChecker.
Initialization of SumatraConversation done in execute().
SystemEnvironment updated to be a non-blocking function when necessary.
