+++ src/nl/hannahsten/texifyidea/run/sumatra/SumatraConversation.kt
@@ -1,93 +1,10 @@
 package nl.hannahsten.texifyidea.run.sumatra
 
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.SystemInfo
 import com.pretty_tools.dde.client.DDEClientConversation
 import nl.hannahsten.texifyidea.TeXception
 import nl.hannahsten.texifyidea.run.linuxpdfviewer.ViewerConversation
-import nl.hannahsten.texifyidea.util.Log
-import nl.hannahsten.texifyidea.util.runCommandWithoutReturn
 import nl.hannahsten.texifyidea.util.runCommandWithExitCode
-import nl.hannahsten.texifyidea.util.runCommand
-import java.io.File
-
-/**
- * Indicates whether SumatraPDF is installed and DDE communication is enabled.
- *
- * Is computed once at initialization (for performance), which means that the IDE needs to be restarted when users
- * install SumatraPDF while running TeXiFy.
- */
-object SumatraAvailabilityChecker {
-
-    private var isSumatraAvailable: Boolean = false
-
-    private var sumatraWorkingCustomDir: File? = null
-
-    private val isSumatraAvailableInit: Boolean by lazy {
-        if (!SystemInfo.isWindows || !isSumatraInstalled()) return@lazy false
-
-        // Try if native bindings are available
-        try {
-            DDEClientConversation()
-        }
-        catch (e: UnsatisfiedLinkError) {
-            Log.info("Native library DLLs could not be found.")
-            return@lazy false
-        }
-        catch (e: NoClassDefFoundError) {
-            Log.info("Native library DLLs could not be found.")
-            return@lazy false
-        }
-
-        true
-    }
-
-    init {
-        isSumatraAvailable = isSumatraAvailableInit
-    }
-
-    fun getSumatraAvailability(): Boolean {
-        return isSumatraAvailable
-    }
-
-    fun getSumatraWorkingCustomDir(): File? {
-        return sumatraWorkingCustomDir
-    }
-
-    /**
-     * Checks if Sumatra can be found in a global PATH or in a directory (with sumatraCustomPath)
-     * Verifies that sumatraCustomPath is a directory, non-null and non-empty before checking in the directory for Sumatra.
-     */
-    fun isSumatraPathAvailable(sumatraCustomPath: String? = null, assignNewAvailability: Boolean = true): Pair<Boolean, File?> {
-        var workingDir: File? = null
-        if (!sumatraCustomPath.isNullOrEmpty() && File(sumatraCustomPath).isDirectory) {
-            workingDir = File(sumatraCustomPath)
-        }
-
-        val availabilityParams = Pair(runCommandWithExitCode("where", "SumatraPDF", workingDirectory = workingDir).second == 0, workingDir)
-
-        if (assignNewAvailability && !isSumatraAvailableInit) {
-            isSumatraAvailable = availabilityParams.first
-            if (isSumatraAvailable && workingDir != null) {
-                sumatraWorkingCustomDir = workingDir
-            }
-        }
-
-        return availabilityParams
-    }
-
-    private fun isSumatraInstalled(): Boolean {
-        // Try some SumatraPDF registry keys
-        // For some reason this first one isn't always present anymore, it used to be
-        val regQuery1 = runCommand("reg", "query", "HKEY_LOCAL_MACHINE\\SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\App Paths\\SumatraPDF.exe", "/ve")?.startsWith("ERROR:") == false
-        val regQuery2 = runCommand("reg", "query", "HKEY_LOCAL_MACHINE\\SOFTWARE\\Classes\\SumatraPDF.pdf", "/ve")?.startsWith("ERROR:") == false
-
-        if (regQuery1 || regQuery2) return true
-
-        // Try if Sumatra is in PATH
-        return isSumatraPathAvailable(sumatraCustomPath = null, assignNewAvailability = false).first
-    }
-}
 
 /**
  * Send commands to SumatraPDF.
@@ -104,7 +21,7 @@ object SumatraConversation : ViewerConversation() {
     private var conversation: DDEClientConversation? = null
 
     private fun openConversation() {
-        if (SumatraAvailabilityChecker.getSumatraAvailability()) {
+        if (SumatraAvailabilityChecker.getSumatraAvailability() && conversation == null) {
             try {
                 conversation = DDEClientConversation()
             }
@@ -118,7 +35,6 @@ object SumatraConversation : ViewerConversation() {
      * Open a file in SumatraPDF, starting it if it is not running yet.
      */
     fun openFile(pdfFilePath: String, newWindow: Boolean = false, focus: Boolean = false, forceRefresh: Boolean = false, sumatraPath: String? = null) {
-        openConversation()
         try {
             execute("Open(\"$pdfFilePath\", ${newWindow.bit}, ${focus.bit}, ${forceRefresh.bit})")
         }
@@ -126,7 +42,7 @@ object SumatraConversation : ViewerConversation() {
             // Added checks when sumatraPath doesn't exist (not a directory), so Windows popup error doesn't appear
             val (_, workingDir) = SumatraAvailabilityChecker.isSumatraPathAvailable(sumatraPath)
             if (SumatraAvailabilityChecker.getSumatraAvailability()) {
-                runCommandWithoutReturn("cmd.exe", "/C", "start", "SumatraPDF", "-reuse-instance", pdfFilePath, workingDirectory = workingDir)
+                runCommandWithExitCode("cmd.exe", "/C", "start", "SumatraPDF", "-reuse-instance", pdfFilePath, workingDirectory = workingDir, nonBlocking = true)
             }
         }
     }
@@ -157,6 +73,7 @@ object SumatraConversation : ViewerConversation() {
     }
 
     private fun execute(vararg commands: String) {
+        openConversation()
         try {
             conversation!!.connect(server, topic)
             conversation!!.execute(commands.joinToString(separator = "") { "[$it]" })
